short stats in php module: reduce memory footprint
Previously the call and key were both cached in the statistics for every call.
Now only the first 500 calls are cached with both call and key,
afterwards only the duration and whether it was cached is saved.
The default threshold of 500 can be manually changed with $palava->set(IpcMemcache::CONFIG_SHORT_STAT_THRESHOLD, ...).